+++ pkg/types/chain_metadata.go
@@ -1,0 +1,6 @@
+package types
+
+type ChainMetadata struct {
+	BlockExplorerUrl           string
+	ELDelegationManagerAddress string
+}
