+++ pkg/utils/constants.go
@@ -10,4 +10,5 @@ const (
 	GoerliChainId  = 5
 	MainnetChainId = 1
 	HoleskyChainId = 17000
+	LocalChainId   = 31337
 )
